fix: greener builds
When an ubuntu build fails, it's very likely that macos & win32 builds
will also fail.

Both of those run slower and cost more. While the price itself it not a
deal breaker, the failure notification is also delayed.

In terms of pricing, by one estimate,
    win32 builds: ~2x(time)   & ~4x(cost)
    macos builds: ~1.7x(time) & ~16x(cost)

This is double the advertized cost for win & macos

To optimize, we split the build into two sequntial jobs,
    - the first job builds & tests on ubuntu
    - the second job builds & tests on win & macos.

At the slight cost of delaying successful builds, it improves the speed
of failure notification while save a few bucks.

Not to mention that since we use fewer CPU minutes overall, we are a
greener build!